pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = ""
authors = ["Ryo Kuroyanagi <im.ryokuro@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.14.0"
python-dotenv = "^1.0.0"
psutil = "^5.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
